refactor(harness): share validation failure helpers
(cherry picked from commit 4ef5e0b34cff391aa0a9773e1a4d891f6df726a2)
